+++ parsers/src/Text/Parser/Char.hs
@@ -1,4 +1,3 @@
-{-# LANGUAGE CPP #-}
 {-# LANGUAGE DefaultSignatures #-}
 {-# LANGUAGE FlexibleContexts #-}
 {-# LANGUAGE ScopedTypeVariables #-}
@@ -60,15 +59,6 @@ import qualified Data.Text as Text
 import Data.Text (Text)
 import qualified Text.ParserCombinators.ReadP as ReadP
 import Text.Parser.Combinators
-
-#ifdef MIN_VERSION_parsec
-import qualified Text.Parsec as Parsec
-#endif
-
-#ifdef MIN_VERSION_attoparsec
-import qualified Data.Attoparsec.Types as Att
-import qualified Data.Attoparsec.Combinator as Att
-#endif
 
 -- | @oneOf cs@ succeeds if the current character is in the supplied
 -- list of characters @cs@. Returns the parsed character. See also
@@ -334,21 +324,6 @@ instance (CharParsing m, MonadPlus m) => CharParsing (IdentityT m) where
   text = lift . text
   {-# INLINE text #-}
 
-#ifdef MIN_VERSION_parsec
-instance Parsec.Stream s m Char => CharParsing (Parsec.ParsecT s u m) where
-  satisfy   = Parsec.satisfy
-  char      = Parsec.char
-  notChar c = Parsec.satisfy (/= c)
-  anyChar   = Parsec.anyChar
-  string    = Parsec.string
-#endif
-
-#ifdef MIN_VERSION_attoparsec
-instance Att.Chunk t => CharParsing (Att.Parser t) where
-  satisfy p = fmap e2c $ Att.satisfyElem $ p . e2c
-    where e2c = Att.chunkElemToChar (undefined :: t)
-  {-# INLINE satisfy #-}
-#endif
 
 instance CharParsing ReadP.ReadP where
   satisfy   = ReadP.satisfy