+++ parsers/src/Text/Parser/Token.hs
@@ -1,4 +1,3 @@
-{-# LANGUAGE CPP #-}
 {-# LANGUAGE FlexibleContexts #-}
 {-# LANGUAGE FlexibleInstances #-}
 {-# LANGUAGE GeneralizedNewtypeDeriving #-}
@@ -113,13 +112,6 @@ import Text.Parser.Char
 import Text.Parser.Combinators
 import Text.Parser.Token.Highlight
 
-#ifdef MIN_VERSION_parsec
-import qualified Text.Parsec as Parsec
-#endif
-
-#ifdef MIN_VERSION_attoparsec
-import qualified Data.Attoparsec.Types as Att
-#endif
 
 -- | Skip zero or more bytes worth of white space. More complex parsers are
 -- free to consider comments as white space.
@@ -927,12 +919,5 @@ instance TokenParsing m => TokenParsing (Unlined m) where
   highlight h (Unlined m) = Unlined (highlight h m)
   {-# INLINE highlight #-}
 
-#ifdef MIN_VERSION_parsec
-instance Parsec.Stream s m Char => TokenParsing (Parsec.ParsecT s u m)
-#endif
-
-#ifdef MIN_VERSION_attoparsec
-instance Att.Chunk t => TokenParsing (Att.Parser t)
-#endif
 
 instance TokenParsing ReadP.ReadP